Split interface file into smaller modules
