+++ src/device/rdk/connectivity.rs
@@ -1,0 +1,3 @@
+pub mod ws;
+pub mod http;
+pub mod rdk;
